init: start postgresql-typeorm task
Refactor functions & add data
