+++ .env
@@ -3,8 +3,8 @@ PORT=4000
 AUTH_MODE=false
 JWT_SECRET_KEY=secret-key
 # Set your database connection information here
-MONGO_CONNECTION_STRING=your-mongo-db-connection-string
+POSTGRES_DBHOST=tai.db.elephantsql.com
 POSTGRES_PORT=5432
-POSTGRES_DB=rsschool
-POSTGRES_USER=rsschool
-POSTGRES_PASSWORD=qwerty123456
+POSTGRES_DB=yddsnhcm
+POSTGRES_USER=yddsnhcm
+POSTGRES_PASSWORD=DmZhyr8yvo7GbGEOTgCouefZ9Rp5agTQ
