+++ src/resources/boards/board.memory.repository.ts
@@ -1,44 +1,47 @@
-import {
-  getAllBoards,
-  getBoard,
-  createBoard,
-  updateBoard,
-  deleteBoard,
-  deleteBoardTasks,
- } from '../../common/database';
- import { IBoardUpdatedBody } from './board.types';
- import { Messages } from '../../common/statusCodes';
+ import { IBoardUpdatedBody, IBoard } from './board.types';
+ import { BOARDS, TASKS } from '../../common/data';
+ import { Board } from './board.model';
 
-const getAll = async () => getAllBoards();
-const create = async (board: IBoardUpdatedBody) => createBoard(board);
-const get = async (id: string) => {
-  const board = await getBoard(id);
-  if (!board) {
-    throw new Error(Messages.BAD_REQUEST);
-  }
-  return board
+const getAll = async () => BOARDS;
+
+const create = async (board: IBoardUpdatedBody): Promise<IBoard> => {
+  const { title, columns } = board;
+  const createdBoard = new Board({ title, columns });
+  BOARDS.push(createdBoard);
+  return createdBoard
 };
-const update = async (id: string, body:IBoardUpdatedBody) => {
-  const board = await updateBoard(id, body);
-  if (!board) {
-    throw new Error(Messages.BAD_REQUEST);
+
+const get = async (id: string): Promise<IBoard | undefined> =>
+  BOARDS.filter(board => board.id === id)[0];
+
+const update = async (id: string, updateBody: IBoardUpdatedBody): Promise<IBoard | undefined> => {
+  const board = BOARDS.filter(element => element.id === id)[0];
+  if (board) {
+    board.title = updateBody.title;
+    board.columns = updateBody.columns;
+    return board
   }
-  return board
+  return undefined
 };
+
 const delBoard = async (id: string) => {
-  const board = await deleteBoard(id);
-  if (!board) {
-    throw new Error(Messages.BAD_REQUEST);
+  const index = BOARDS.findIndex((board) => board.id === id);
+  if (index < 0) {
+    return null;
   }
-  return board
+  BOARDS.splice(index, 1);
+  return true;
 };
 
-const delTasks = async (id: string) => {
-  const tasks = await deleteBoardTasks(id);
-  if (!tasks) {
-    throw new Error(Messages.BAD_REQUEST);
-  }
-  return tasks
+const delTasks = async (boardId: string) => {
+  TASKS.map((task) => {
+    const copyTask = task;
+    if (copyTask.boardId === boardId) {
+      copyTask.boardId = null;
+    }
+    return copyTask;
+  });
+  return true;
 };
 
 export { getAll, create, get, update, delBoard, delTasks };
