+++ src/resources/tasks/task.memory.repository.ts
@@ -1,35 +1,44 @@
-import {
-  getAllTasks,
-  getTask,
-  createTask,
-  updateTask,
-  deleteTask,
- } from '../../common/database';
  import { ITaskUpdatedBody } from './task.types';
- import { Messages } from '../../common/statusCodes';
+ import { TASKS } from '../../common/data';
+ import { Task } from './task.model';
 
-const getAll = async () => getAllTasks();
-const create = async (task: ITaskUpdatedBody) => createTask(task);
-const get = async (taskId: string, boardId: string) => {
-  const task = await getTask(taskId, boardId);
-  if (!task) {
-    throw new Error(Messages.NOT_FOUND);
-  }
-  return task
+
+const getAll = async () => TASKS;
+
+const create = async (task: ITaskUpdatedBody) => {
+  const { title, order, description, userId, boardId, columnId } = task;
+  const createdTask = new Task({ title, order, description, userId, boardId, columnId });
+  TASKS.push(createdTask);
+  return createdTask;
 };
-const update = async (id: string, boardId: string, body: ITaskUpdatedBody) => {
-  const task = updateTask(id, boardId, body);
-  if (!task) {
-    throw new Error(Messages.NOT_FOUND);
+
+const get = async (taskId: string, boardId: string) => TASKS.find(
+  task => task.id === taskId && task.boardId === boardId
+);
+
+const update = async (id: string, boardId: string, updateBody: ITaskUpdatedBody) => {
+  const task = TASKS.filter(element => element.id === id && element.boardId === boardId)[0];
+  if (task) {
+    task.title = updateBody.title;
+    task.order = updateBody.order;
+    task.description = updateBody.description;
+    task.userId = updateBody.userId;
+    task.boardId = updateBody.boardId;
+    task.columnId = updateBody.columnId;
+    return task
   }
-  return task
+  return undefined
 };
+
 const delTask = async (taskId: string, boardId: string) => {
-  const task = await deleteTask(taskId, boardId);
-  if (!task) {
-    throw new Error(Messages.NOT_FOUND);
+  const index = TASKS.findIndex(
+    (task) => task.boardId === boardId && task.id === taskId
+  );
+  if (index < 0) {
+    return null;
   }
-  return task
+  TASKS.splice(index, 1);
+  return true
 };
 
-module.exports = { getAll, create, get, update, delTask };
+export { getAll, create, get, update, delTask };
